sqs-lambda-cdk/src/requirements.txt
aws-cdk.core
aws-cdk.aws-lambda
aws-cdk.aws-logs
aws-cdk.aws-iam
aws-cdk.aws-sqs
aws-cdk.aws-lambda-event-sources